Convert bools to lowercase for XML spec compliance
`bool.ToString()` returns an uppercase "True" or "False", but XML is
case sensitive and speicifes that bools should be lowercase (see
https://learn.microsoft.com/en-us/dotnet/api/system.boolean.tostring).

While the built-in NuspecReader handles this, tools like NuGet Package
Explorer to not. Update the nuspec writer in Package to emit lowercase
bools, and update unit test to catch accidental regressions.